% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar.R
\name{CYTARDatasource}
\alias{CYTARDatasource}
\title{CYTARDatasource}
\description{
CYTARDatasource

CYTARDatasource
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARDatasource$new()}}
\item \href{#method-loadData}{\code{CYTARDatasource$loadData()}}
\item \href{#method-consolidate}{\code{CYTARDatasource$consolidate()}}
\item \href{#method-clone}{\code{CYTARDatasource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDatasource$new(data.url, data.filename, col.types, delim = ";")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDatasource$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDatasource$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDatasource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
