% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar-personas.R
\name{CYTARDisciplinasRef}
\alias{CYTARDisciplinasRef}
\title{CYTARDisciplinasRef}
\description{
CYTARDisciplinasRef

CYTARDisciplinasRef
}
\author{
kenarab
}
\section{Super class}{
\code{\link[CYTAR:CYTARDatasource]{CYTAR::CYTARDatasource}} -> \code{CYTARDisciplinasRef}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARDisciplinasRef$new()}}
\item \href{#method-checkConsolidatedFields}{\code{CYTARDisciplinasRef$checkConsolidatedFields()}}
\item \href{#method-consolidate}{\code{CYTARDisciplinasRef$consolidate()}}
\item \href{#method-clone}{\code{CYTARDisciplinasRef$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CYTAR" data-topic="CYTARDatasource" data-id="loadData">}\href{../../CYTAR/html/CYTARDatasource.html#method-loadData}{\code{CYTAR::CYTARDatasource$loadData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDisciplinasRef$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkConsolidatedFields"></a>}}
\if{latex}{\out{\hypertarget{method-checkConsolidatedFields}{}}}
\subsection{Method \code{checkConsolidatedFields()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDisciplinasRef$checkConsolidatedFields(fields)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDisciplinasRef$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARDisciplinasRef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
