% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar.R
\name{CYTARPersonas2018}
\alias{CYTARPersonas2018}
\title{CYTARPersonas2018}
\description{
CYTARPersonas2018

CYTARPersonas2018
}
\author{
kenarab
}
\section{Super class}{
\code{\link[CYTAR:CYTARDatasource]{CYTAR::CYTARDatasource}} -> \code{CYTARPersonas2018}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARPersonas2018$new()}}
\item \href{#method-clone}{\code{CYTARPersonas2018$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CYTAR" data-topic="CYTARDatasource" data-id="loadData">}\href{../../CYTAR/html/CYTARDatasource.html#method-loadData}{\code{CYTAR::CYTARDatasource$loadData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARPersonas2018$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARPersonas2018$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
