% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar-personas.R
\name{CYTARPersonasAnioDownloader}
\alias{CYTARPersonasAnioDownloader}
\title{CYTARPersonasAnioDownloader}
\description{
CYTARPersonasAnioDownloader

CYTARPersonasAnioDownloader
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARPersonasAnioDownloader$new()}}
\item \href{#method-configure}{\code{CYTARPersonasAnioDownloader$configure()}}
\item \href{#method-generatePersonasyear}{\code{CYTARPersonasAnioDownloader$generatePersonasyear()}}
\item \href{#method-clone}{\code{CYTARPersonasAnioDownloader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARPersonasAnioDownloader$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configure"></a>}}
\if{latex}{\out{\hypertarget{method-configure}{}}}
\subsection{Method \code{configure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARPersonasAnioDownloader$configure()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generatePersonasyear"></a>}}
\if{latex}{\out{\hypertarget{method-generatePersonasyear}{}}}
\subsection{Method \code{generatePersonasyear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARPersonasAnioDownloader$generatePersonasyear(year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARPersonasAnioDownloader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
