% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar-producciones.R
\name{CYTARProducciones}
\alias{CYTARProducciones}
\title{current.yearTARPersonas2018}
\description{
current.yearTARPersonas2018

current.yearTARPersonas2018
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARProducciones$new()}}
\item \href{#method-configAll}{\code{CYTARProducciones$configAll()}}
\item \href{#method-loadAll}{\code{CYTARProducciones$loadAll()}}
\item \href{#method-clone}{\code{CYTARProducciones$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProducciones$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-configAll"></a>}}
\if{latex}{\out{\hypertarget{method-configAll}{}}}
\subsection{Method \code{configAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProducciones$configAll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadAll"></a>}}
\if{latex}{\out{\hypertarget{method-loadAll}{}}}
\subsection{Method \code{loadAll()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProducciones$loadAll()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProducciones$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
