% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar-producciones.R
\name{CYTARProduccionYear}
\alias{CYTARProduccionYear}
\title{CYTARProduccionYear}
\description{
CYTARProduccionYear

CYTARProduccionYear
}
\author{
kenarab
}
\section{Super class}{
\code{\link[CYTAR:CYTARDatasource]{CYTAR::CYTARDatasource}} -> \code{CYTARProduccionYear}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARProduccionYear$new()}}
\item \href{#method-loadData}{\code{CYTARProduccionYear$loadData()}}
\item \href{#method-clone}{\code{CYTARProduccionYear$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProduccionYear$new(data.url, year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProduccionYear$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProduccionYear$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
