% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar-producciones.R
\name{CYTARProductoPersonaFuncion}
\alias{CYTARProductoPersonaFuncion}
\title{CYTARProductoPersonaFuncionYear}
\description{
CYTARProductoPersonaFuncionYear

CYTARProductoPersonaFuncionYear
}
\author{
kenarab
}
\section{Super class}{
\code{\link[CYTAR:CYTARDatasource]{CYTAR::CYTARDatasource}} -> \code{CYTARProductoPersonaFuncion}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARProductoPersonaFuncion$new()}}
\item \href{#method-clone}{\code{CYTARProductoPersonaFuncion$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CYTAR" data-topic="CYTARDatasource" data-id="consolidate">}\href{../../CYTAR/html/CYTARDatasource.html#method-consolidate}{\code{CYTAR::CYTARDatasource$consolidate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CYTAR" data-topic="CYTARDatasource" data-id="loadData">}\href{../../CYTAR/html/CYTARDatasource.html#method-loadData}{\code{CYTAR::CYTARDatasource$loadData()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoPersonaFuncion$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoPersonaFuncion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
