% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytar-producciones.R
\name{CYTARProductoYear}
\alias{CYTARProductoYear}
\title{CYTARProductoYear}
\description{
CYTARProductoYear

CYTARProductoYear
}
\author{
kenarab
}
\section{Super class}{
\code{\link[CYTAR:CYTARDatasource]{CYTAR::CYTARDatasource}} -> \code{CYTARProductoYear}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CYTARProductoYear$new()}}
\item \href{#method-loadData}{\code{CYTARProductoYear$loadData()}}
\item \href{#method-processKeywords}{\code{CYTARProductoYear$processKeywords()}}
\item \href{#method-getProducciones}{\code{CYTARProductoYear$getProducciones()}}
\item \href{#method-clone}{\code{CYTARProductoYear$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoYear$new(data.url, year)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoYear$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-processKeywords"></a>}}
\if{latex}{\out{\hypertarget{method-processKeywords}{}}}
\subsection{Method \code{processKeywords()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoYear$processKeywords()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProducciones"></a>}}
\if{latex}{\out{\hypertarget{method-getProducciones}{}}}
\subsection{Method \code{getProducciones()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoYear$getProducciones(producciones.id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CYTARProductoYear$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
